+++ src/test/resources/source_support/Foo.java
@@ -1,0 +1,7 @@
+public class Foo {
+
+    public static void main(String[] args) {
+        Bar.test(2);
+    }
+
+}