+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.7] - 2022-08-29
+
+### Added
+
+- Support for Java source file via Soot's source parser. This can be disabled via 
+`Jimple2Cpg.createCpg(parseJavaSource)`.
+
 ## [1.2.6] - 2022-05-31
 
 ### Changed